requirements.txt
aiogram==2.22.1
async-lru==1.0.2
async-timeout==4.0.2
flake8==3.9.2
flake8-docstrings==1.6.0
pytest==6.2.5
python-dotenv==0.19.0
python-telegram-bot==13.7
requests==2.26.0